Add common module with utils
